requirements.txt
confluent-kafka
pandas
psycopg2-binary
elasticsearch
neo4j
fastapi
uvicorn
requests
